Update database credentials and adjust Catering list display
Changes include updating the login credentials in the application resources. Username reference in the Catering Lists title on the front-end is also updated to match the appropriate value. Additionally, the test for CateringListService now checks whether the returned list from findByUserId() is not empty instead of checking the user's id for validation. Several minor code clean-ups have also been done in the UserRepository file.